♻️ refactor(v0.5.2-elements-split): split converter/elements.py into a package
Mechanical move of the 1519-line monolith into 4 modules + facade
(bodies unchanged; only imports rewired to the parent package):

- shapes.py   name maps + line / freeform / auto-shape extraction
- textbox.py  textbox extraction (delegates preset geometry to shapes)
- media.py    video / picture / SVG / image-fill / referenced images
- dispatch.py raw passthrough, type routing, group recursion
- __init__.py compatibility facade (7 public extractors + _dispatch_shape)

Dependency DAG (textbox->shapes, dispatch->{shapes,textbox,media}) is
enforced by an AST guard; submodule import smoke tests included.
Characterization expectations unchanged — 615 passed.

SPEC: 20260731-2310_v0.5.2-elements-split
Progress: Phase 3 (package split) complete
Next: Phase 4 warning closure
